+++ Chunk.cpp
@@ -21,14 +21,14 @@ void ChunkPrefab::GenerateChunkSurface()
 {
 	for (int x = 0; x < this->xSize; x++) {
 		for (int z = 0; z < this->zSize; z++) {
-			int Height = (int)(35 + ( PerlinNoise({ (double)xPos + x, 0, (double)zPos + z }, 4, 0.1f) * 25));
+			int Height = (int)(35 + ( PerlinNoise({ (float)xPos + x, 0, (float)zPos + z }, 4, 0.1f) * 25));
 			int ActualHeight = Height;
 			if (Height < 35) {
 				Height = 35;
 			}
 
 			for (int y = Height; y > -1; y--) {
-				Vector3 BlockPos = { (double)x, (double)y, (double)z };
+				Vector3 BlockPos = { (float)x, (float)y, (float)z };
 				if (y <= ActualHeight) {
 					Blocks[BlockPos] = 3;
 					/*
@@ -61,12 +61,12 @@ void ChunkPrefab::GenerateChunkCaves()
 		for (int z = 0; z < this->xSize; z++) {
 			for (int y = 2; y < this->ySize; y++)
 			{
-				Vector3 BlockPos = { (double)x, (double)y, (double)z };
+				Vector3 BlockPos = { (float)x, (float)y, (float)z };
 				auto it = Blocks.find(BlockPos);
 				if (it == Blocks.end()) continue;
 				int blockID = it->second;
 
-				double Hole = PerlinNoise({ (double)xPos + x, (double)y , (double) z}, 3, 0.1f);
+				float Hole = PerlinNoise({ (float)xPos + x, (float)y , (float) z}, 3, 0.1f);
 
 				bool CheeseCave = Hole <= -0.9f || Hole >= 0.9f;
 				bool NodleCave = (0.04f > Hole && Hole > -0.04f);
@@ -85,7 +85,7 @@ void ChunkPrefab::VisableFaces() {
 	for (int y = 0; y < this->ySize; y++) {
 		for (int x = 0; x < this->xSize; x++) {
 			for (int z = 0; z < this->zSize; z++) {
-				Vector3 blockPos = { (double)x, (double)y, (double)z };
+				Vector3 blockPos = { (float)x, (float)y, (float)z };
 
 				auto it = Blocks.find(blockPos);
 				if (it == Blocks.end()) continue;
@@ -97,7 +97,7 @@ void ChunkPrefab::VisableFaces() {
 					auto blockIt = Blocks.find(NextBlockPos);
 
 					if (blockIt == Blocks.end() || (blockID != 5 && blockIt->second == 5)) {
-						Vector3 ChunkWorldPos = { (double)this->xPos, 0, (double)this->zPos };
+						Vector3 ChunkWorldPos = { (float)this->xPos, 0, (float)this->zPos };
 						Vector3 world = blockPos + ChunkWorldPos;
 						this->allFaces.push_back({ world, i, blockID, 0 });
 					}
